Replace the OOSEQ max bytes and pbufs constants with functions.
Some systems need to take into account an RX buffer pool size when
advising an appropriate number of RX pbufs to queue on the ooseq
list. For some systems there is a practical hard limit beyond which
the rx pool becomes exhausted blocking reception of further buffers
until some are freed.

It also helps to be able to consider the available dynamic memory when
advising an appropriate maximum number of bytes to buffer on the ooseq
list.

These decisions can also benefit from knowing the number already
allocated on a particular pcb, so this is passed to these
functions. For example, if the system only wants to allow the total
number of rx pbufs queued on all the ooseq lists to grow by one and a
call is make indicating a pcb already has two then it can return three
for this call, but might return one for another call - supporting a
greedy allocation strategy.