mdns resp: separate the announce function and add netifapi support for it.
The mDNS responder uses a lot of stack for a small system. This patch
attempts to mitigate this by allowing the announce code paths to run
in the tcpip thread where the stack can be provisioned, and removing
that burden from all threads that might invoke an announcement.

The interface setup functions no longer invoke an announcement. Typically a
series are used to setup the state and it is only useful to send an announcement
at the end anyway. The caller is now responsible to invoking an announcement as
needed after making changes.

Netifapi support is added for invoking an announcement. This was trivial to add
as it takes only the netif as an argument. It is left to future work to add
other api support for the mDNS responder.

The strategy only works when core locking is not used, otherwise the code paths
can be invoked in other threads that make interface changes.