sys_timeout: correct insertion of new timeouts.
When checking if the new timeout should be inserted at the head of the
next_timeout list, consistently compare the differences to the
timeouts_last_time rather that comparing a difference to the timeout
in msec from now.

Remove logic described as catching a large difference between
timeouts_last_time and now(). This could have been triggered by a
small delay in processing the timers. Just insert in the appropriate
order.