mdns: move data off the stack and dynamically allocate data.
The mdns responder uses a lot of stack for a small system, well over
1k, which is out of proportion to the general stack resource
needs. This is largely due to stack allocated fixed sized 256 byte
buffers for domain names. This patch largely moves data off the stack
and dynamically allocates it.

The code also allocated fixed size buffers for the domain names, big
enough for the maximum size name of 256 bytes. This patch dynamically
allocates buffers just large enough for their content, growing them as
needed. There are trade offs in doing so, but with this change it is
still trivial to over allocate storage if that is preferred.