sys_check_timeouts: recalculate the time diff after each handler call
If there is only one timeout then when the handler is called there
will be none. If the handler then calls sys_timeout to install another
handler and then sys_timeout will reset the timeouts_last_time. When
the handler returns to the sys_check_timeouts loop it was continuing
to use a stale diff value and incorrectly checking the times.

This patch modifies sys_check_timeouts to sample sys_now() and
timeouts_last_time each time around the loop. This allowed some
cleanup of the code.

Also PBUF_CHECK_FREE_OOSEQ() is now called even if there are no timeouts.