lwip_cyclic_timer: keep consistent trigger periods in the long run.
When a timer would immediate re-trigger after a handler returns, the
current code adjusts the trigger time into the future. This
unfortunately leads to uneven time periods and does not appear to be
necessary, so simply allow it to re-trigger immediately. This change
provides a consistent timer period in the long run, even if there is
some short term jitter.