mDNS: heap allocate rather than stack allocate.
The mdns domains with a default size of 256 were being stack allocated and
this used a lot of stack space. These are now dynamically allocated from the
heap. Also only the required length is allocated rather than the static
maximum size, so the memory usage will typically be less, although
fragmentation might be an issue.

A few other stack and static allocations are now also dynamically allocated
from the heap.

This strategy does not appear to be compatible with the lwip memory allocation
approach which is to work without malloc, but it helps on low memory devices.

This is a rebase of these changes to this significantly modified mDNS code
base.